chore(src): remove dead code
