--- src/SlateCommands/ConvertToHeaderChange.js
@@ -1,3 +1,0 @@
-export default function ConvertToHeaderChange(editor) {
-  editor.setBlocks('heading');
-}
