--- src/SlateCommands/ConvertToParagraphChange.js
@@ -1,3 +1,0 @@
-export default function ConvertToParagraphChange(editor) {
-  editor.setBlocks('paragraph');
-}
