--- src/SlateCommands/ToggleBlockTypeChange.js
@@ -1,4 +1,0 @@
-export default function ToggleBlockTypeChange(editor, type) {
-  const isType = editor.value.blocks.some(block => block.type === type);
-  editor.setBlocks(isType ? 'paragraph' : type);
-}
