+++ src/plugins/withClauseSchema.js
@@ -1,6 +1,7 @@
 export const CLAUSE = 'clause';
 export const VARIABLE = 'variable';
 export const CONDITIONAL = 'conditional';
+export const COMPUTED = 'computed';
 
 const INLINES = { [VARIABLE]: true, [CONDITIONAL]: true };
 const VOIDS = {};
